url_shortner_backend/requirement.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
PyJWT==2.10.1
redis==5.2.1
sqlparse==0.5.3
tzdata==2024.2
